📝 ContextValues (move annotation registration)
