+++ src/main/java/io/github/syst3ms/skriptparser/event/ScriptLoadContext.java
@@ -1,5 +1,6 @@
 package io.github.syst3ms.skriptparser.event;
 
+import io.github.syst3ms.skriptparser.Parser;
 import io.github.syst3ms.skriptparser.lang.TriggerContext;
 import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValueMethod;
 
@@ -8,6 +9,10 @@ import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValueMet
  * in Java.
  */
 public class ScriptLoadContext implements TriggerContext {
+    static {
+        Parser.getMainRegistration().addContextAnnotations(ScriptLoadContext.class);
+    }
+
     private final String[] args;
 
     public ScriptLoadContext(String[] args) {
