+++ src/main/java/io/github/syst3ms/skriptparser/registration/SkriptRegistration.java
@@ -27,6 +27,7 @@ import io.github.syst3ms.skriptparser.pattern.TextElement;
 import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValue;
 import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValue.State;
 import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValue.Usage;
+import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValueMethod;
 import io.github.syst3ms.skriptparser.registration.contextvalues.ContextValues;
 import io.github.syst3ms.skriptparser.registration.tags.Tag;
 import io.github.syst3ms.skriptparser.registration.tags.TagInfo;
@@ -41,6 +42,7 @@ import io.github.syst3ms.skriptparser.util.MultiMap;
 import org.jetbrains.annotations.Nullable;
 
 import java.lang.reflect.Array;
+import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -385,6 +387,52 @@ public class SkriptRegistration {
     }
 
     /**
+     * Registers all {@linkplain ContextValue}s generated by the
+     * context annotations for methods and fields.
+     * @param context the TriggerContext classes
+     * @param <C> the TriggerContext class
+     * @see ContextValueMethod
+     */
+    @SuppressWarnings("unchecked")
+    public <C extends TriggerContext> void addContextAnnotations(Class<C> context) {
+        for (var method : context.getMethods()) {
+            if (method.getAnnotation(ContextValueMethod.class) != null
+                    && method.getParameterTypes().length == 0) {
+                // Less initialisation
+                var annotation = method.getAnnotation(ContextValueMethod.class);
+                var returnType = method.getReturnType();
+                var type = TypeManager.getByClass((Class<?>) (returnType.isArray() ? returnType.getComponentType() : returnType));
+                var pattern = PatternParser.parsePattern(annotation.name(), logger);
+
+                if (type.isEmpty()) {
+                    logger.error("Couldn't find a type corresponding to the class '" + returnType.getName() + "'", ErrorType.NO_MATCH);
+                    return;
+                } else if (pattern.isEmpty()) {
+                    return;
+                }
+
+                // Now all conditions have been satisfied
+                contextValues.add(new ContextValue<>(
+                        registerer,
+                        context,
+                        (Type<Object>) type.get(),
+                        !returnType.isArray(),
+                        pattern.get(),
+                        value -> {
+                            try {
+                                return returnType.isArray() ? (Object[]) method.invoke(value) : new Object[] {method.invoke(value)};
+                            } catch (IllegalAccessException | InvocationTargetException ex) {
+                                throw new IllegalStateException("ContextParameter method could not be invoked without any parameters");
+                            }
+                        },
+                        annotation.state(),
+                        annotation.usage()
+                ));
+            }
+        }
+    }
+
+    /**
      * Starts a registration process for a {@link Type}
      * @param c the class the Type represents
      * @param pattern the Type's pattern
@@ -660,31 +708,31 @@ public class SkriptRegistration {
 
         /**
          * Starts a registration process for a {@link ContextExpression}
-         * @param context the context this value appears in
+         * @param context the TriggerContext class
          * @param returnType the returned type of this context value
          * @param isSingle whether or not the return value is single
          * @param pattern the pattern
          * @param function the function that needs to be applied in order to get the context value
-         * @param <C> the ContextValue
+         * @param <C> the TriggerContext class
          * @param <R> the ContextValue's return type
          * @return a {@link ContextValueRegistrar} to continue the registration process
          */
-        public final <C extends TriggerContext, R> ContextValueRegistrar<C, R> newContextValue(Class<C> context, Class<R> returnType, boolean isSingle, String pattern, Function<C, R[]> function) {
+        public <C extends TriggerContext, R> ContextValueRegistrar<C, R> newContextValue(Class<C> context, Class<R> returnType, boolean isSingle, String pattern, Function<C, R[]> function) {
             return new ContextValueRegistrar<>(context, returnType, isSingle, pattern, function);
         }
 
         /**
          * Registers a {@link ContextValue}
-         * @param context the context this value appears in
+         * @param context the TriggerContext class
          * @param returnType the returned type of this context value
          * @param isSingle whether or not the return value is single
          * @param pattern the pattern
          * @param function the function that needs to be applied in order to get the context value
-         * @param <C> the ContextValue
+         * @param <C> the TriggerContext class
          * @param <R> the ContextValue's return type
          * @return this {@link EventRegistrar}
          */
-        public final <C extends TriggerContext, R> EventRegistrar<T> addContextValue(Class<C> context, Class<R> returnType, boolean isSingle, String pattern, Function<C, R[]> function) {
+        public <C extends TriggerContext, R> EventRegistrar<T> addContextValue(Class<C> context, Class<R> returnType, boolean isSingle, String pattern, Function<C, R[]> function) {
             new ContextValueRegistrar<>(context, returnType, isSingle, pattern, function).register();
             return this;
         }
@@ -693,14 +741,14 @@ public class SkriptRegistration {
          * Registers a {@link ContextValue} that returns a single value.
          * The {@linkplain Type#getBaseName() base name} of the return type will be used as pattern.
          * There will be a leading '{@code [the] }' in the pattern if the context value can be used alone.
-         * @param context the context this value appears in
+         * @param context the TriggerContext class
          * @param returnType the returned type of this context value
          * @param function the function that needs to be applied in order to get the context value
-         * @param <C> the ContextValue
+         * @param <C> the TriggerContext class
          * @param <R> the ContextValue's return type
          * @return this {@link EventRegistrar}
          */
-        public final <C extends TriggerContext, R> EventRegistrar<T> addContextType(Class<C> context, Class<R> returnType, Function<C, R> function) {
+        public <C extends TriggerContext, R> EventRegistrar<T> addContextType(Class<C> context, Class<R> returnType, Function<C, R> function) {
             return addContextType(context, returnType, function, State.PRESENT, Usage.EXPRESSION_ONLY);
         }
 
@@ -708,17 +756,17 @@ public class SkriptRegistration {
          * Registers a {@link ContextValue} that returns a single value.
          * The {@linkplain Type#getBaseName() base name} of the return type will be used as pattern.
          * There will be a leading '{@code [the] }' in the pattern if the context value can be used alone.
-         * @param context the context this value appears in
+         * @param context the TriggerContext class
          * @param returnType the returned type of this context value
          * @param function the function that needs to be applied in order to get the context value
          * @param state the time state
-         * @param <C> the ContextValue
+         * @param <C> the TriggerContext class
          * @param <R> the ContextValue's return type
          * @return this {@link EventRegistrar}
          * @see State#PRESENT
          * @see Usage#EXPRESSION_ONLY
          */
-        public final <C extends TriggerContext, R> EventRegistrar<T> addContextType(Class<C> context, Class<R> returnType, Function<C, R> function, State state) {
+        public <C extends TriggerContext, R> EventRegistrar<T> addContextType(Class<C> context, Class<R> returnType, Function<C, R> function, State state) {
             return addContextType(context, returnType, function, state, Usage.EXPRESSION_ONLY);
         }
 
@@ -726,24 +774,24 @@ public class SkriptRegistration {
          * Registers a {@link ContextValue} that returns a single value.
          * The {@linkplain Type#getBaseName() base name} of the return type will be used as pattern.
          * There will be a leading '{@code [the] }' in the pattern if the context value can be used alone.
-         * @param context the context this value appears in
+         * @param context the TriggerContext class
          * @param returnType the returned type of this context value
          * @param function the function that needs to be applied in order to get the context value
          * @param state the time state
          * @param usage the usage
-         * @param <C> the ContextValue
+         * @param <C> the TriggerContext class
          * @param <R> the ContextValue's return type
          * @return this {@link EventRegistrar}
          * @see Usage#EXPRESSION_ONLY
          */
         @SuppressWarnings("unchecked")
-        public final <C extends TriggerContext, R> EventRegistrar<T> addContextType(Class<C> context, Class<R> returnType, Function<C, R> function, State state, Usage usage) {
+        public <C extends TriggerContext, R> EventRegistrar<T> addContextType(Class<C> context, Class<R> returnType, Function<C, R> function, State state, Usage usage) {
             var typeName = TypeManager.getByClassExact(returnType).map(Type::getBaseName).orElse(TypeManager.NULL_REPRESENTATION);
             new ContextValueRegistrar<>(context, returnType, true, typeName, value -> {
-                        R[] array = (R[]) Array.newInstance(returnType, 1);
-                        array[0] = function.apply(value);
-                        return array;
-                    }).setState(state).setUsage(usage).register();
+                R[] array = (R[]) Array.newInstance(returnType, 1);
+                array[0] = function.apply(value);
+                return array;
+            }).setState(state).setUsage(usage).register();
             return this;
         }
